requirements.txt
mysql-connector-python
flask
beautifulsoup4
requests
flask-cors